pyproject.toml
[tool.poetry]
name = "vms-fatmug"
version = "0.1.0"
description = ""
authors = ["Aman Rawat <ar.rawataman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
drf-nested-routers = "^0.93.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
